Component, System 코드 정리 Async 객체들의 Ref 카운팅 작업중 Recv, Send,  Timer에 추가할 예정

+++ Server/Base/Thread.cpp
@@ -47,7 +47,7 @@ void CThread::Run()
 
 		Milli_t waitTick = Math::Min(CalculateWaitTick() - duration , (Milli_t)0);
 
-		Sleep(waitTick);
+		Sleep((DWORD)waitTick);
 	}
 }
